[Feat #339] ModalConfetti.ts 생성
